refactor(viewer): share tool badge rendering
Reuse the same context and duration badge components across tool renderers and keep Cursor JSONL result types near their parser usage.

Made-with: Cursor
