+++ packages/viewer/src/components/Badges.tsx
@@ -3,6 +3,8 @@
  * styling stays consistent — change the pill once, all variants update.
  */
 
+import { formatToolDuration, formatTokens } from "./StatsPanel";
+
 export function ErrorBadge() {
   return (
     <span
@@ -10,6 +12,32 @@ export function ErrorBadge() {
       title="This tool call returned an error"
     >
       ERROR
+    </span>
+  );
+}
+
+export function ToolDuration({ ms }: { ms?: number }) {
+  const label = formatToolDuration(ms);
+  if (!label) return null;
+  return (
+    <span
+      className="text-[10px] text-terminal-dimmer font-mono shrink-0"
+      title={`Tool execution: ${label}`}
+    >
+      {label}
+    </span>
+  );
+}
+
+export function ToolTokens({ tokens }: { tokens?: number }) {
+  const label = formatTokens(tokens);
+  if (!label) return null;
+  return (
+    <span
+      className="text-[10px] text-terminal-dimmer font-mono shrink-0"
+      title={`~${label} tokens added to context (heuristic estimate)`}
+    >
+      ~{label} tok
     </span>
   );
 }
