+++ packages/viewer/src/components/CodeDiffBlock.tsx
@@ -1,8 +1,7 @@
 import { diffLines as computeLineDiff } from "diff";
 import { Highlight, themes } from "prism-react-renderer";
 import { memo, useMemo, useSyncExternalStore } from "react";
-import { ErrorBadge } from "./Badges";
-import { formatToolDuration, formatTokens } from "./StatsPanel";
+import { ErrorBadge, ToolDuration, ToolTokens } from "./Badges";
 
 // Subscribe to dark/light class changes on <html> for prism theme switching
 const subscribe = (cb: () => void) => {
@@ -88,8 +87,6 @@ export default memo(function CodeDiffBlock({
   const isDark = useSyncExternalStore(subscribe, getIsDark);
   const isNewFile = !oldContent;
   const isDeletedFile = toolName === "Delete" && !newContent;
-  const tokenLabel = formatTokens(resultTokens);
-  const durationLabel = formatToolDuration(durationMs);
 
   return (
     <div
@@ -105,22 +102,8 @@ export default memo(function CodeDiffBlock({
         </span>
         <span className="text-xs font-mono text-terminal-blue truncate flex-1">{filePath}</span>
         {isError && <ErrorBadge />}
-        {tokenLabel && (
-          <span
-            className="text-[10px] text-terminal-dimmer font-mono shrink-0"
-            title={`~${tokenLabel} tokens added to context`}
-          >
-            ~{tokenLabel} tok
-          </span>
-        )}
-        {durationLabel && (
-          <span
-            className="text-[10px] text-terminal-dimmer font-mono shrink-0"
-            title={`Execution: ${durationLabel}`}
-          >
-            {durationLabel}
-          </span>
-        )}
+        <ToolTokens tokens={resultTokens} />
+        <ToolDuration ms={durationMs} />
         {isNewFile && (
           <span className="text-xs px-1.5 py-0.5 rounded bg-terminal-green/20 text-terminal-green">
             new
